Add useSearchField
