+++ packages/peregrine/lib/talons/SearchBar/index.js
@@ -1,2 +1,3 @@
 export { useAutocomplete } from './useAutocomplete';
 export { useSearchBar } from './useSearchBar';
+export { useSearchField } from './useSearchField';
